修改

+++ camerafragment/src/main/java/com/github/florent37/camerafragment/internal/utils/CameraHelper.java
@@ -114,7 +114,7 @@ public final class CameraHelper {
             }
         }
 
-        if (mediaFile.exists()) {
+        if (mediaFile != null && mediaFile.exists()) {
             mediaFile.delete();
         }
